as1750: Split types.h from type.h, clean up integer types for portability across 32/64 bit hosts

+++ as1750/utils.h
@@ -22,7 +22,7 @@ extern char  *findc (char *s, char c);
 extern char  *skip_white (char *s);
 extern char  *skip_nonwhite (char *s);
 extern char  *skip_symbol (char *s);
-extern long  get_16bit_hexnum (char *s);
-extern long  get_nibbles (char *src, int n_nibbles);
-extern void  put_nibbles (char *dst, unsigned long src, int n_nibbles);
+extern int   get_16bit_hexnum (char *s);
+extern int   get_nibbles (char *src, int n_nibbles);
+extern void  put_nibbles (char *dst, unsigned int src, int n_nibbles);
